Restructuring tests
setting up a pytest structure. Updating old CAMx readers to
fix out of date testing
