+++ src/PseudoNetCDF/camxfiles/irr/Read.py
@@ -307,8 +307,7 @@ class TestRead(unittest.TestCase):
         pass
         
     def testIRR(self):
-        emissfile=irr('../../../../testdata/ei/camx_cb4_ei_lo.20000825.hgb8h.base1b.psito2n2.hgbpa_04km')
-        self.assert_(1==2)
-       
+        warn('Test not implemented; data too big for distribution')
+        
 if __name__ == '__main__':
     unittest.main()
