+++ src/PseudoNetCDF/camxfiles/one3d/Read.py
@@ -125,7 +125,7 @@ class one3d(PseudoNetCDFFile):
         self.start_time,self.start_date=self.rffile.read(self.id_fmt)
         self.record_size=self.rffile.record_size
         self.padded_size=self.record_size+8
-        self.cell_count=(self.record_size-self.id_size)/struct.calcsize(self.data_fmt)
+        self.cell_count=(self.record_size-self.id_size)//struct.calcsize(self.data_fmt)
         self.record_fmt=self.id_fmt+self.data_fmt*(self.cell_count)
         
     def __gettimestep(self):
@@ -246,7 +246,7 @@ class one3d(PseudoNetCDFFile):
         a=zeros((self.time_step_count,self.nlayers,len(self.dimensions['ROW']),len(self.dimensions['COL'])),'f')
         for ti,(d,t) in enumerate(self.timerange()):
             for ki,k in enumerate(range(1,self.nlayers+1)):
-                self.seekandreadinto(a[ti,ki,...,...],d,t,k)
+                self.seekandreadinto(a[ti,ki,...],d,t,k)
         return a
 
     def timerange(self):
@@ -259,14 +259,8 @@ class TestRead(unittest.TestCase):
         pass
         
     def testKV(self):
-        vdfile=one3d('../../../../testdata/met/camx_kv.20000825.hgbpa_04km.TCEQuh1_eta.v43.tke',65,83)
-        self.assert_((vdfile.variables['UNKNOWN'].mean(0).mean(1).mean(1)==array([  13.65080357,   34.39198303,   68.02783966,   95.5898819 ,
-          109.25765991,  112.92014313,  108.32209778,   97.25794983,
-          84.1328125 ,   65.92033386,   46.97774506,   25.8343792 ,
-          9.80327034,    2.89653206,    1.26993668,    1.12098336,
-          1.13557184,    1.13372564,    1.19559622,    1.1675849 ,
-          1.18877947,    1.18713808,    1.02371764,    1.02544105,
-          1.21638143,    1.34624374,    1.03213251,    1.        ],dtype='f')).all())
-       
+        import PseudoNetCDF.testcase
+        vdfile=one3d(PseudoNetCDF.testcase.camxfiles_paths['vertical_diffusivity'],4,5)
+        self.assert_((vdfile.variables['UNKNOWN']==array([1.00000000e+00, 4.76359320e+00, 1.92715893e+01, 1.52158489e+01, 7.20601225e+00, 1.84097159e+00, 2.63084507e+01, 1.27621298e+01, 1.06348248e+01, 2.22587357e+01, 1.00000000e+00, 1.69009724e+01, 1.00000000e+00, 2.23075104e+01, 1.27485418e+01, 1.45508013e+01, 1.45637455e+01, 2.95294094e+01, 3.02676849e+01, 2.84974957e+01, 1.00000000e+00, 6.93706131e+00, 3.07418957e+01, 3.41300621e+01, 1.66266994e+01, 1.00000000e+00, 2.53920174e+01, 1.92539787e+01, 2.32906532e+01, 5.96702042e+01, 1.00000000e+00, 2.24458847e+01, 1.00000000e+00, 5.45038452e+01, 3.45825729e+01, 3.43578224e+01, 3.76071548e+01, 6.76799850e+01, 7.33648529e+01, 7.35801239e+01, 1.00000000e+00, 6.38178444e+00, 4.39327278e+01, 5.00754166e+01, 2.44474106e+01, 1.00000000e+00, 3.51700935e+01, 2.71137428e+01, 3.40312347e+01, 8.85775909e+01, 1.00000000e+00, 3.13994522e+01, 1.00000000e+00, 8.07169266e+01, 5.12892876e+01, 5.05734329e+01, 5.56603966e+01, 1.00394188e+02, 1.08980370e+02, 1.09251083e+02, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.64916098e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.22205174e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.05408611e+01, 1.26687222e+01, 1.21386652e+01, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.93040049e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.39350021e+00, 1.02697349e+00, 1.00000000e+00, 1.00000000e+00, 1.82250175e+01, 2.90407104e+01, 2.83827496e+01, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 2.02609706e+00, 1.00000000e+00, 1.00000000e+00, 1.00000000e+00, 1.44422662e+00, 1.02998519e+00, 1.00000000e+00, 1.00000000e+00, 2.60322971e+01, 4.26534195e+01, 4.17046585e+01],dtype='f').reshape(2,3,4,5)).all())       
 if __name__ == '__main__':
     unittest.main()
