+++ src/PseudoNetCDF/camxfiles/temperature/Read.py
@@ -114,7 +114,7 @@ class temperature(PseudoNetCDFFile):
         self.rffile._newrecord(0)
         
         self.area_size=self.rffile.record_size
-        self.area_count=(self.area_size-self.id_size)/struct.calcsize(self.data_fmt)
+        self.area_count=(self.area_size-self.id_size)//struct.calcsize(self.data_fmt)
         self.area_padded_size=self.area_size+8
         self.area_fmt=self.id_fmt+self.data_fmt*(self.area_count)
 
@@ -122,7 +122,7 @@ class temperature(PseudoNetCDFFile):
         
         self.record_size=self.rffile.record_size
         self.padded_size=self.record_size+8
-        self.cell_count=(self.record_size-self.id_size)/struct.calcsize(self.data_fmt)
+        self.cell_count=(self.record_size-self.id_size)//struct.calcsize(self.data_fmt)
         
         self.record_fmt=self.id_fmt+self.data_fmt*(self.cell_count)
     
@@ -136,7 +136,7 @@ class temperature(PseudoNetCDFFile):
         self.rffile.infile.seek(0,2)
         self.rffile.previous()
         self.end_time,self.end_date=self.rffile.read(self.id_fmt)
-        self.time_step_count=int(timediff((self.start_date,self.start_time),(self.end_date,self.end_time))/self.time_step)+1
+        self.time_step_count=int(timediff((self.start_date,self.start_time),(self.end_date,self.end_time))//self.time_step)+1
     
     def __variables(self,k):
         if k=='SURFTEMP':
@@ -145,8 +145,9 @@ class temperature(PseudoNetCDFFile):
         elif k=='AIRTEMP':
             out=zeros((len(self.dimensions['TSTEP']),len(self.dimensions['LAY']),len(self.dimensions['ROW']),len(self.dimensions['COL'])),'f')
             vars=self.__airmaps()
-        for i,(d,t) in enumerate(self.timerange()):
-            out[i,...]=vars.next()
+        dts = self.timerange()
+        for i, v in enumerate(vars):
+            out[i,...]=v
         return out
         
     def __surfpos(self):
@@ -187,16 +188,9 @@ class TestRead(unittest.TestCase):
         pass
     
     def testTEMP(self):
-        vdfile=temperature('../../../../testdata/met/camx_temp.20000825.hgbpa_04km.TCEQuh1_eta.v43',65,83)
-        self.assert_((vdfile.variables['AIRTEMP'].mean(0).mean(1).mean(1)==array([300.78424072265625, 
-               300.6566162109375, 300.28036499023438, 299.61105346679688, 298.90249633789062, 
-               298.17489624023438, 297.45626831054688, 296.75238037109375, 296.08837890625, 
-               295.47381591796875, 294.83306884765625, 294.19754028320312, 293.681396484375, 
-               293.14077758789062, 292.3377685546875, 291.16769409179688, 289.93109130859375, 
-               288.52767944335938, 286.92172241210938, 284.93429565429688, 282.81634521484375, 
-               279.27914428710938, 274.59872436523438, 269.8642578125, 263.99749755859375, 
-               252.75810241699219, 230.50096130371094, 209.34048461914062],dtype='f')).all())
-    
+        import PseudoNetCDF.testcase
+        tempfile=temperature(PseudoNetCDF.testcase.camxfiles_paths['temperature'],4,5)
+        self.assert_((tempfile.variables['AIRTEMP']==array([2.97762360e+02, 2.97261993e+02, 3.00761200e+02, 3.03811005e+02, 3.04561218e+02, 2.96350311e+02, 2.96676544e+02, 3.00992096e+02, 3.05474762e+02, 3.07840637e+02, 2.99522430e+02, 3.00271698e+02, 3.03738403e+02, 3.07201843e+02, 3.08288422e+02, 3.02957214e+02, 3.04927643e+02, 3.06630157e+02, 3.07726074e+02, 3.07380707e+02, 2.97516449e+02, 2.96920105e+02, 3.00340576e+02, 3.03413177e+02, 3.04202728e+02, 2.96074036e+02, 2.96250641e+02, 3.00632294e+02, 3.05113647e+02, 3.07390533e+02, 2.99310059e+02, 2.99901031e+02, 3.03344666e+02, 3.06782135e+02, 3.07819946e+02, 3.02657013e+02, 3.04522675e+02, 3.06167206e+02, 3.07235107e+02, 3.06883484e+02, 2.97677338e+02, 2.96919098e+02, 3.00031250e+02, 3.03082672e+02, 3.03850861e+02, 2.96460999e+02, 2.95947815e+02, 3.00303680e+02, 3.04781982e+02, 3.07048492e+02, 2.99246979e+02, 2.99508667e+02, 3.02997650e+02, 3.06450500e+02, 3.07478485e+02, 3.02246765e+02, 3.04192139e+02, 3.05832489e+02, 3.06897644e+02, 3.06546173e+02, 2.97428253e+02, 2.97174896e+02, 3.00208191e+02, 3.03096893e+02, 3.04174133e+02, 2.96558685e+02, 2.96706177e+02, 3.00862610e+02, 3.04807037e+02, 3.06937347e+02, 2.98850220e+02, 2.99482727e+02, 3.03085022e+02, 3.06456787e+02, 3.07406586e+02, 3.01888580e+02, 3.03996735e+02, 3.05916962e+02, 3.07113647e+02, 3.06539337e+02, 2.97645966e+02, 2.97326630e+02, 3.00117950e+02, 3.02804077e+02, 3.03801544e+02, 2.96783661e+02, 2.96694946e+02, 3.00722931e+02, 3.04501587e+02, 3.06560150e+02, 2.98854828e+02, 2.99314972e+02, 3.02861023e+02, 3.06150177e+02, 3.07073944e+02, 3.01700745e+02, 3.03746124e+02, 3.05626617e+02, 3.06770447e+02, 3.06172394e+02, 2.97927094e+02, 2.97691681e+02, 3.00104675e+02, 3.02464874e+02, 3.03398926e+02, 2.97336578e+02, 2.97074127e+02, 3.00716736e+02, 3.04132446e+02, 3.06129700e+02, 2.98817017e+02, 2.99221039e+02, 3.02649536e+02, 3.05787415e+02, 3.06698334e+02, 3.01333618e+02, 3.03411346e+02, 3.05317505e+02, 3.06446869e+02, 3.05815948e+02],dtype='f').reshape(2,3,4,5)).all())    
         
 if __name__ == '__main__':
     unittest.main()
