+++ src/PseudoNetCDF/test/test_geoschemfiles.py
@@ -1,0 +1,3 @@
+from PseudoNetCDF.geoschemfiles._bpch import TestMemmaps as test_bpch
+from PseudoNetCDF.geoschemfiles._newbpch import TestMemmaps as test_bpch2
+from PseudoNetCDF.geoschemfiles._geos import TestMemmaps as test_geos
