+++ src/PseudoNetCDF/test/test_icarttfiles.py
@@ -1,0 +1,1 @@
+from PseudoNetCDF.icarttfiles.ffi1001 import TestFfi1001 as test_ffi1001
