+++ src/PseudoNetCDF/test/test_pncopen.py
@@ -1,0 +1,1 @@
+from PseudoNetCDF._getreader import TestPNCOPEN as test_pncopen
